# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Xpart App OpenApi API Documentation",
        "description": "Xpart App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "showOrders",
                "description": "showOrders",
                "operationId": "showOrders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order Receipt number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "orders",
                "description": "orders",
                "operationId": "orders",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "postOrders",
                "description": "postOrders",
                "operationId": "postOrders",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "showOrders",
                "description": "showOrders",
                "operationId": "showOrders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "updateOrders",
                "description": "updateOrders",
                "operationId": "updateOrders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Addresses ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderCreateFormRequest": {
                "title": "Order Create Form Request Fields",
                "description": "Order Create Form request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "address_id": {
                        "title": "User Address id",
                        "description": "Address ID of the user",
                        "type": "integer",
                        "example": "1"
                    },
                    "cart": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "itemable_id": {
                                    "type": "int",
                                    "example": "1"
                                },
                                "itemable_type": {
                                    "type": "string",
                                    "example": "quotes"
                                },
                                "quantity": {
                                    "type": "int",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "orderable_id": {
                        "title": "Orderable id",
                        "description": "Address ID of this transaction",
                        "type": "integer",
                        "example": "1"
                    },
                    "orderable_type": {
                        "title": "Orderable type",
                        "description": "Orderable type of this transaction",
                        "type": "integer",
                        "example": "1"
                    },
                    "payment_method_id": {
                        "title": "Payment Method id",
                        "description": "Payment Method ID of this transaction",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "OrderUpdateFormRequest": {
                "title": "Order Update Form Request Fields",
                "description": "Order Update Form request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "payment_reference": {
                        "title": "Order payment reference",
                        "description": "Order payment reference",
                        "type": "string",
                        "example": "quote"
                    },
                    "payment_method_id": {
                        "title": "Order payment method id",
                        "description": "Order payment method id",
                        "type": "string",
                        "example": "quote"
                    },
                    "payment_gateway": {
                        "title": "Order payment gateway",
                        "description": "Order payment gateway",
                        "type": "string",
                        "example": "quote"
                    },
                    "order_id": {
                        "title": "Order ID ",
                        "description": "Order Id ",
                        "type": "integer",
                        "example": "quote"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}